+++ src/main/java/org/folio/rest/client/PasswordResetActionClient.java
@@ -4,6 +4,7 @@ import io.vertx.core.Future;
 import org.folio.rest.jaxrs.model.PasswordRestAction;
 import org.folio.rest.util.OkapiConnectionParams;
 
+import java.util.Optional;
 
 /**
  * Client for password reset actions API in mod-login module
@@ -18,4 +19,13 @@ public interface PasswordResetActionClient {
    * @return future with true if password already exists
    */
   Future<Boolean> saveAction(PasswordRestAction passwordRestAction, OkapiConnectionParams okapiConnectionParams);
+
+  /**
+   * Retrieves an password reset action with given id
+   *
+   * @param passwordResetActionId password reset action id
+   * @param okapiConnectionParams connection params
+   * @return
+   */
+  Future<Optional<PasswordRestAction>> getAction(String passwordResetActionId, OkapiConnectionParams okapiConnectionParams);
 }