+++ src/main/java/org/folio/service/PasswordResetLinkService.java
@@ -1,6 +1,7 @@
 package org.folio.service;
 
 import io.vertx.core.Future;
+import org.folio.rest.jaxrs.model.TokenResponse;
 import org.folio.rest.util.OkapiConnectionParams;
 
 /**
@@ -15,4 +16,13 @@ public interface PasswordResetLinkService {
    * @return Future with generated link
    */
   Future<String> sendPasswordRestLink(String userId, OkapiConnectionParams okapiConnectionParams);
+
+  /**
+   * Retrieves reset password action id from x-okapi-token connection param, retrieves password reset action,
+   * validates the action, then signs new system JWT token.
+   *
+   * @param okapiConnectionParams connection params
+   * @return new JWT token
+   */
+  Future<TokenResponse> validateLinkAndLoginUser(OkapiConnectionParams okapiConnectionParams);
 }